Updates Made in this commit: .) Edit Screen now updates ingredient name / type changes on Shared Data and its own Screen
Next To Do:
.) Edit Screen delete / refresh btn

Previous Updates:
.)

Errors
.)

+++ src/main/java/App_Code/Objects/Screens/Ingredient_Info_Screens/Ingredients_Info/Add_Ingredients/Ingredients_Form.java
@@ -600,7 +600,7 @@ public class Ingredients_Form extends Parent_Forms_OBJ
             Field_JTxtField_BD salt_Obj = (Field_JTxtField_BD) field_Items_Map.get("salt").get_Gui_Component();
             
             // Convert mg to G
-            BigDecimal salt_MG_Value = (BigDecimal) salt_Obj.get_Text_Casted_To_Type();
+            BigDecimal salt_MG_Value = salt_Obj.get_Text_Casted_To_Type();
             
             // Convert MG to G value
             BigDecimal new_Value_G = salt_MG_Value.divide(new BigDecimal("1000"), 2, RoundingMode.HALF_UP);
